[ndk][crystax] Revert "[X86] Part 2 to fix x86-64 fp128 calling convention."
This reverts commit 3a1999a31131b95b9ae19cdcfb8bc00bf585ed85.

This is done in order to fix bug https://tracker.crystax.net/issues/1363

For some reason, this commit cause clang to crash when compiling simple
code with long double for x86_64. It's not clear why this happens;
however, with this commit reverted everything seems to be working.
So revert this commit for a while and get back to it later.

Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>